섹션 6-2 : Composed annotation을 만들어서 적용해보기
