+++ src/main/kotlin/com/example/sbt/annotation/MySpringBootAnnotation.kt
@@ -1,0 +1,12 @@
+package com.example.sbt.annotation
+
+import org.springframework.boot.SpringBootConfiguration
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.CLASS)
+@Configuration
+@ComponentScan
+@SpringBootConfiguration
+annotation class MySpringBootAnnotation()
