+++ src/main/kotlin/com/example/sbt/config/Config.kt
@@ -1,0 +1,18 @@
+package com.example.sbt.config
+
+import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.DispatcherServlet
+
+@Configuration
+class Config {
+
+    @Bean
+    fun servletWebServerFactory(): ServletWebServerFactory = TomcatServletWebServerFactory()
+
+    @Bean
+    fun dispatcherServlet() = DispatcherServlet()
+
+}
